requirements.txt
# ===== DEPENDENCIAS ESENCIALES =====
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-CORS==6.0.1
Flask-Compress==1.18
Werkzeug==3.1.3

# ===== INTELIGENCIA ARTIFICIAL =====
google-generativeai==0.8.5
google-auth==2.40.3

# ===== VERTEX AI (MIGRACIÓN RECOMENDADA) =====
google-cloud-aiplatform==1.70.0
vertexai==1.70.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1

# ===== BASE DE DATOS =====
SQLAlchemy==2.0.43
psycopg2-binary

# ===== AUTENTICACIÓN Y SEGURIDAD =====
PyJWT==2.9.0
Flask-JWT-Extended==4.6.0
bcrypt==4.3.0
cryptography==43.0.1

# ===== SSL/HTTPS SOPORTE =====
pyOpenSSL==24.2.1
certifi==2025.8.3

# ===== VALIDACIÓN =====
WTForms==3.1.2
email-validator==2.2.0

# ===== CONFIGURACIÓN =====
python-dotenv==1.0.1
click==8.2.1

# ===== PROCESAMIENTO DE ARCHIVOS =====
Pillow==11.3.0
PyPDF2==3.0.1
PyMuPDF==1.26.3
pytesseract==0.3.10
python-magic==0.4.27
opencv-python==4.9.0.80
numpy
scipy

# ===== UTILIDADES =====
requests==2.32.4

# ===== DEPENDENCIAS DE DESARROLLO =====
pytest==8.2.0
pytest-flask==1.3.0
pytest-cov==4.1.0
Flask-Testing==0.8.1

# ===== VALIDACIÓN Y SEGURIDAD =====
bleach==6.1.0

# ===== MONITOREO Y MÉTRICAS =====
psutil==5.9.6

# ===== PRODUCCIÓN =====
gunicorn==23.0.0